requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
joblib==1.3.2
python-multipart==0.0.9
pydantic==2.6.3
requests==2.31.0
dvc==3.49.0
openai==1.14.3
python-dotenv==1.0.1
openpyxl==3.1.2
